# Improve LRU Get speed