--- a/storelru.go
+++ b/storelru.go
@@ -56,29 +56,18 @@ func (s lruStore[K, V]) Set(key K, value Item[K, V]) {
 }
 
 func (s lruStore[K, V]) Get(key K) (Item[K, V], bool) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
+	s.mu.RLock()
 	item, ok := s.items[key]
 	if !ok {
 		return Item[K, V]{}, false
 	}
+	s.mu.RUnlock()
 
+	s.mu.Lock()
 	s.list.MoveToFront(s.elements[key])
+	s.mu.Unlock()
 
 	return item, true
-
-	// TODO: compare against this:
-	// s.mu.RLock()
-	// item, ok := s.items[key]
-	// if !ok {
-	// 	return Item[K, V]{}, false
-	// }
-	// s.mu.RUnlock()
-	// s.mu.Lock()
-	// s.list.MoveToFront(s.elements[key])
-	// s.mu.Unlock()
-	// return item, true
 }
 
 func (s lruStore[K, V]) Items() (map[K]Item[K, V], unlockFunc) {