Before:
```
go test -benchmem -run=^$ -coverprofile=/tmp/vscode/go-code-cover -bench ^BenchmarkCache_Get_parallel$ github.com/wafer-bw/memcache -benchtime=100000000x
goos: linux
goarch: amd64
pkg: github.com/wafer-bw/memcache
cpu: AMD EPYC 7763 64-Core Processor                
BenchmarkCache_Get_parallel/noevict_policy_parallel-2           100000000               26.29 ns/op            0 B/op          0 allocs/op
BenchmarkCache_Get_parallel/lru_policy_parallel-2               100000000               42.20 ns/op            0 B/op          0 allocs/op
PASS
coverage: 30.9% of statements
ok      github.com/wafer-bw/memcache    6.856s
```

After
```
go test -benchmem -run=^$ -coverprofile=/tmp/vscode/go-code-cover -bench ^BenchmarkCache_Get_parallel$ github.com/wafer-bw/memcache -benchtime=100000000x
goos: linux
goarch: amd64
pkg: github.com/wafer-bw/memcache
cpu: AMD EPYC 7763 64-Core Processor                
BenchmarkCache_Get_parallel/noevict_policy_parallel-2           100000000               26.26 ns/op            0 B/op          0 allocs/op
BenchmarkCache_Get_parallel/lru_policy_parallel-2               100000000               37.76 ns/op            0 B/op          0 allocs/op
PASS
coverage: 32.0% of statements
ok      github.com/wafer-bw/memcache    6.409s
```